created product_view, created one more event to inputs to calc unit values

+++ database/migrations/2020_04_22_030008_create_control_storages_table.php
@@ -15,7 +15,7 @@ class CreateControlStoragesTable extends Migration
     {
       Schema::create('control_storages', function (Blueprint $table) {
           $table->bigIncrements('id');
-          $table->bigInteger('quantidade');
+          $table->float('quantidade');
           $table->string('unidade_venda', 10);
           $table->unsignedBigInteger('produto_id');
           $table->foreign('produto_id')
